block page added and profile page impreved and code arranged

+++ client/src/components/header/Header.js
@@ -1,9 +1,12 @@
 import "./header.css";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import Logo from "../../images/ColorMediaLogo.jpg";
+import Swal from "sweetalert2"
 import default_profile from "../../images/default_profile.png"
+// import { io } from "socket.io-client";
+
+
 import {
-  ArrowDown,
   Menu,
   Messenger,
   Notifications,
@@ -11,15 +14,25 @@ import {
 } from "../../svg";
 import { useDispatch, useSelector } from "react-redux";
 import useInstance from "../../axios/axiosInstance.js";
-
+import { useState } from "react";
 
  const Header = ({page}) => {
+  const navigate = useNavigate()
   let instance = useInstance()
+  const [searchText,setSearchText] = useState("")
   const { user } = useSelector((user) => ({ ...user }));
-  console.log(user);
   const dispatch = useDispatch()
   
   const color = "#65676b";
+
+
+//   const socket = io("ws://localhost:5000");
+// socket.emit("welcome",{data:"hello from client"})
+// socket.on("welcome",(data)=>{
+//   console.log("Message from server",data);
+// })
+ 
+
   return (
     <header>
       <div className="header_left">
@@ -34,6 +47,16 @@ import useInstance from "../../axios/axiosInstance.js";
             type="text"
             placeholder="Search...."
             className="hide_input"
+            value={searchText}
+            onKeyDown={(e)=>{
+                if(e.key==="Enter"){
+                  navigate(`/profile/${searchText}`)
+                }
+            }}
+            onChange={(e)=>{
+              setSearchText(e.target.value)
+            }}
+          
           />
         </div>
       </div>
@@ -62,28 +85,47 @@ import useInstance from "../../axios/axiosInstance.js";
       <div className="header_right">
       <Link to="/profile" className={`profile_link hover1 ${page === "profile" ? "active_link" : ""}`}>
           <img src={user?.picture ||default_profile} alt="Profile" />
-          <span>{user?.first_name}</span>
+          <span>{user?.username}</span>
         </Link>
         <div className="circle_icon hover1">
           <Menu />
         </div>
-        <div className="circle_icon hover1">
+        <div  className={`circle_icon hover1 ${page === "message" ? "active_link" : ""}`} onClick={()=>{
+            navigate("/message")
+        }}>
           <Messenger />
-        <div className="right_notification">99+</div>          
+        <div className={`${page === "message" ? "" : "right_notification"}`}>{`${page === "message" ? "" : "99+"}`}</div>          
         </div>
         <div className="circle_icon hover1">
           <Notifications />
           <div className="right_notification">8</div>
         </div>
-        <div className="circle_icon hover1"  onClick={async ()=>{
-
-           await instance.patch("/logout")
-            localStorage.setItem("user",null)
-            dispatch({
-              type:"LOGOUT"
-            })
-        }}>
-          <ArrowDown />
+        
+        <div className="circle_icon hover1"  onClick={
+          ()=>{
+      
+          Swal.fire({
+            title: 'Are you sure to logout ?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Logout'
+          }).then(async(result) => {
+            if (result.isConfirmed) {
+                await instance.patch("/logout")
+                 localStorage.setItem("user",null)
+                 dispatch({
+                   type:"LOGOUT"
+                 })
+            }
+          })
+                 
+          }
+          
+         }>
+          <i className="fa-solid fa-right-from-bracket"/>
         </div>
       </div>
     </header>
